refactor(plugin-host): deliver slot host controls via prop for cross-bundle safety
Slot contributions received their host controls through a React context, which
does not cross a separately-bundled plugin's module boundary. Deliver the value
as a prop instead, matching how runtimeContext is already passed. Move
SetupHostContextValue/SetupCredentials to types.ts and remove the provider/hook.
