+++ src/plugin-host/types.ts
@@ -83,6 +83,32 @@ export interface PluginRuntimeContextDescriptor {
   useValue: () => RuntimeContextValue;
 }
 
+/** Generic credentials the host hands a post-bootstrap step (e.g. to upload
+ *  assets behind an authenticated endpoint). Not brand-specific. */
+export interface SetupCredentials {
+  email: string;
+  password: string;
+}
+
+export interface SetupHostContextValue {
+  /** True while the host's bootstrap call is in flight. */
+  isLoading: boolean;
+  /** Host-level error string ('' when none). */
+  error: string;
+  /** Return to the account step. */
+  goBack: () => void;
+  /**
+   * Finish the install. The host merges `extensionPayload` into the single
+   * /setup/bootstrap request under the opaque `extension_payload` key, and on
+   * success invokes `afterBootstrap` with generic credentials. The host assigns
+   * NO meaning to `extensionPayload`.
+   */
+  submit: (
+    extensionPayload?: Record<string, unknown>,
+    afterBootstrap?: (credentials: SetupCredentials) => Promise<void>,
+  ) => void;
+}
+
 export interface PluginManifest {
   id: string;
   onBoot?: () => void;
